Always use correct tool version
Use go dependency management to download and run common build tools like
controller-gen, ginkgo and kustomize. This will ensure that the correct
version of the tool is used. As a bonus, we can use tools like
dependabot to automatically track and propose updated to our tools.

Under the hood this works by using a custom go module with a single
dependency (the tool). `go run` is used to load and execute the tool at
runtime for that module. Golang caches the compiled version of the tool
and invalidates the cache when a new version of the tool is desired.

This PR does not change how envtest is setup, which is a bit more
complicated.

This PR also aligns the module name with the git repo.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>