Create injector package
The injector is the core of the service binding implementation. It
applies a ServiceBinding to a Workload as mapped by a
ClusterWorkloadResourceMapping. This library is able to be called
from a controller, webhook or cli client.

A custom ClusterWorkloadResourceMapping is looked up for the workload.
If no mapping is defined a mapping appropriate for a PodSpecable
resource is used. Internally, the workload is reflected into a
MetaPodTemplate, the binding applied, and then reflected back to the
workload.

The specific service binding rules remain to be implemented. Each
section of the spec should be independently covered by a future PR.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>