Use kapp to manage deploys from source
`kapp` is a Carvel tool that manages k8s resources. Like kubectl, kapp
is a client side cli that operates on vanilla k8s yaml.

kapp will:
- track resources it creates
- provide a summary of the changes it is about to apply
- optionally, diff the current and new state
- optionally, prompt to confirm the changeset
- watch resource for them to become ready. (i.e. Deployment rollouts are
  complete and new Pods are ready)
- prune fields no longer set (unlike `kubectl apply` which merges)
- delete unused resources (unlike `kubectl apply` which orphens)

Interally, kapp uses ConfigMaps to track which resources it is
responsible for managing. These ConfigMaps need to be created in a
namespace. We default this to kube-system, but can be overridden by
setting KAPP_NS. Additional options can be set by overridding KAPP_ARGS.

https://carvel.dev/kapp/

Signed-off-by: Scott Andrews <scott@andrews.me>