Implement Workload Projection
Implements the full Workload Projection section of the Service Binding
Spec. The behavior is applied within the projector package meaning it is
implicitly compatible with Workload Resource Mappings.

The service binding can be projected and unprojected from the workload.
Additionally, as the ServiceBinding resource is updated, the workload is
also updated to remove the previous projection and apply the updated
projection.

Custom type and provider values are projected using the Downward API
that references an annotation on the Pod that holds the custom value.
The custom values are supported in both the volume and env vars.

In addition to annotations being used to as the source of values to the
Downward API, they are also used to link individual values to the
projected volumes, volumemounts and envvars.

The volumes, volumemounts and envvars arrays are sorted to ensure the
order ServiceBinding resources targeting the same Workload does not
matter. The order of existing values are not modified.

Signed-off-by: Scott Andrews <scott@andrews.me>