Centralize API resolution logic
There are three types of things that we need to be able to dynamically
resolve at runtime:
1. workloads
2. binding secrets (via services)
3. workload mappings

Each of is slightly more complicated than your standard client.Get call
as we need to work with dynamic types, and apply behavior from the spec.

The MappingSource previously in the projector, has moved into this new
package. The new Resolver interface is a superset of the projector's
desired MappingSource.

Signed-off-by: Scott Andrews <scott@andrews.me>