Implicitly track resources loaded in the resolver
We can implicitly track all resources loaded via .Get() in the resolver
without exposing the reconciler-runtime config as part of the contract.
This is done by wrapping the config object, intercepting the calls to
.Get() and then delegating those calls to .TrackAndGet().

This wrapper is only applied within the controller as the tracks are not
useful outside of that context.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>