Split ClusterWorkloadResourceMappings into a separate file
Attempting to use kubectl to apply a CRD and a resource of the new kind
creates a race condition for the API Server to reify the new CRD before
the resource instance is created.

Instead we can split these resources into a separate config file that
can be independently applied to the cluster.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>