Bump github.com/vmware-labs/reconciler-runtime to 0.8.0
Most of this diff is whitespace caused by reformating the test suites
from a list to a map. The new style improves code folding in editors.

It also takes advantage of new capabilities to streamline a couple
corner cases in the tests and removes code that is no longer required.

There are no semantic changes.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>