Limit local and ci build to arm64 and amd64
We now limit the architectures to build images for to linux/arm64 and
linux/amd64 for local and ci builds. All platforms are still built for
tags.

The specific platforms to build can still be overridden using the
KO_PLATFORMS envvar setting the value `all` or a comma separate list of
`${GOOS}/${GOARCH}`. For example `linux/amd64,linux/arm64`.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>