Update k8s, controller-runtime, reconciler-runtime and dies (#227)
Breaking changes in k8s and controller-runtime require updating each
component in lock step.

The only change that directly affects us is that the fake api client now
properly supports list requests using a field index. Previously, the
response would ignore the field index and return all resources. The
request is rejected unless the index is defined on the fake client, so
we now define the index for test cases that use field indexed requests.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>