Default KO_PLATFORMS to only the current arch when kind is the target
ko has special behavior for a `KO_DOCKER_REPO` value of `kind.local`, it
will load the images directly into the daemon rather than a registry.
Kind clusters do not support multi-arch images, we should assume by
default that the kind cluster is running on the same arch as the host.

A user may continue to override the default by setting `KO_PLATFORMS`
with the value they want to use at runtime.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>