Sample using External Secrets Operator
External Secrets Operator (ESO) is a way to synthesize Kubernetes
Secrets from secure data stores. There is a PR that adds provisioned
service support to ExternalSecret resources.

This sample is intended to facilitate the conversation between the
Service Bindings and External Secrets communities. This PR should not be
merged unless and until the update PR lands.

Refs https://github.com/external-secrets/external-secrets/pull/2263

Signed-off-by: Scott Andrews <andrewssc@vmware.com>