Add lifecycle hooks
The hooks allow defining a custom Resolver or Projector as well as
interacting with the ServiceBinding and workload resources before before
and after projection.

Defining no hooks is equivalent to the current RI behavior.

The controller operates on a single ServiceBinding and zero-to-many
workloads in a single request, while the webhook operates on a single
workload with zero-to-many ServiceBindings. Because of this mismatch, we
split out interactions with the ServiceBinding and workload resources
into independent hooks.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>