Keep kube-rbac-proxy images up to date
Kubebuilder includes a kube-rbac-proxy sidecar with the controller
deployment, but provides no mechanism to keep that image up to date.

We now use dependabot to consume new upstream releases of the proxy go
module, and build that source into an image as part of `make deploy` and
the regular CI release via `ko`.

Signed-off-by: Scott Andrews <scott@andrews.me>