Deduplicate Admission Webhooks (#497)
Defining admission webhooks for each version of the same resource
results in duplicate webhook calls for the same admission request. The
resource is converted to each version as that version's webhook is
called. Fortunately these calls were made in parallel, so there was
minimal latency added for users, but the extra calls are just wasting
resources.

We now only register admission webhooks for the v1 version of the
resources. API requests using the v1beta1 or v1alpha3 versions are
converted to the v1 by the API server and then called using the v1
webhook endpoint.

Signed-off-by: Scott Andrews <scott@andrews.me>